# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Looks N Styles Beauty Salon API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Service {
  id: number;
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  category: string;
  /** @nullable */
  imageUrl: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface CreateServiceBody {
  name: string;
  description: string;
  price: number;
  durationMinutes: number;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface UpdateServiceBody {
  name?: string;
  description?: string;
  price?: number;
  durationMinutes?: number;
  category?: string;
  /** @nullable */
  imageUrl?: string | null;
  isActive?: boolean;
}

export interface Appointment {
  id: number;
  customerName: string;
  customerPhone: string;
  serviceId: number;
  serviceName: string;
  appointmentDate: string;
  timeSlot: string;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface CreateAppointmentBody {
  customerName: string;
  customerPhone: string;
  serviceId: number;
  appointmentDate: string;
  timeSlot: string;
  /** @nullable */
  notes?: string | null;
}

export interface UpdateAppointmentBody {
  status?: string;
  /** @nullable */
  notes?: string | null;
}

export interface TimeSlot {
  time: string;
  available: boolean;
}

export interface AvailableSlotsResponse {
  date: string;
  slots: TimeSlot[];
}

export interface AppointmentStats {
  total: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  todayCount: number;
}

export interface GalleryImage {
  id: number;
  imageUrl: string;
  /** @nullable */
  caption: string | null;
  /** @nullable */
  category: string | null;
  sortOrder: number;
  createdAt: string;
}

export interface CreateGalleryImageBody {
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  /** @nullable */
  category?: string | null;
  sortOrder?: number;
}

export interface Review {
  id: number;
  customerName: string;
  rating: number;
  comment: string;
  isApproved: boolean;
  createdAt: string;
}

export interface CreateReviewBody {
  customerName: string;
  rating: number;
  comment: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone: string | null;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export interface CreateContactMessageBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
}

export interface AdminLoginBody {
  password: string;
}

export interface AdminLoginResponse {
  success: boolean;
  message: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
}

export type GetAvailableSlotsParams = {
  date: string;
  serviceId: number;
};

export type ListReviewsParams = {
  approved?: string;
};