# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Looks N Styles Beauty Salon API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: services
    description: Salon services
  - name: appointments
    description: Appointment bookings
  - name: gallery
    description: Gallery images
  - name: reviews
    description: Customer reviews
  - name: contact
    description: Contact messages
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List all services
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateServiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get a service by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateServiceBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
        "404":
          description: Not found
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments (admin)
      responses:
        "200":
          description: List of appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"
    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "409":
          description: Time slot already booked

  /appointments/available-slots:
    get:
      operationId: getAvailableSlots
      tags: [appointments]
      summary: Get available time slots for a date and service
      parameters:
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: serviceId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Available time slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailableSlotsResponse"

  /appointments/stats:
    get:
      operationId: getAppointmentStats
      tags: [appointments]
      summary: Get appointment statistics (admin)
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentStats"

  /appointments/{id}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get an appointment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAppointmentBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
        "404":
          description: Not found
    delete:
      operationId: deleteAppointment
      tags: [appointments]
      summary: Delete appointment (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery images
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"
    post:
      operationId: createGalleryImage
      tags: [gallery]
      summary: Add gallery image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGalleryImageBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryImage"

  /gallery/{id}:
    delete:
      operationId: deleteGalleryImage
      tags: [gallery]
      summary: Delete gallery image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List approved reviews
      parameters:
        - name: approved
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Submit a review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewBody"
      responses:
        "201":
          description: Created (pending approval)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/{id}/approve:
    patch:
      operationId: approveReview
      tags: [reviews]
      summary: Approve a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  /reviews/{id}:
    delete:
      operationId: deleteReview
      tags: [reviews]
      summary: Delete a review (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /contact:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List contact messages (admin)
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
    post:
      operationId: createContactMessage
      tags: [contact]
      summary: Send a contact message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateContactMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid credentials

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminMeResponse"
        "401":
          description: Not authenticated

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Service:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - price
        - durationMinutes
        - category
        - imageUrl
        - isActive
        - createdAt

    CreateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
      required:
        - name
        - description
        - price
        - durationMinutes
        - category

    UpdateServiceBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        durationMinutes:
          type: integer
        category:
          type: string
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean

    Appointment:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        serviceId:
          type: integer
        serviceName:
          type: string
        appointmentDate:
          type: string
        timeSlot:
          type: string
        status:
          type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - customerName
        - customerPhone
        - serviceId
        - serviceName
        - appointmentDate
        - timeSlot
        - status
        - notes
        - createdAt

    CreateAppointmentBody:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        serviceId:
          type: integer
        appointmentDate:
          type: string
        timeSlot:
          type: string
        notes:
          type: ["string", "null"]
      required:
        - customerName
        - customerPhone
        - serviceId
        - appointmentDate
        - timeSlot

    UpdateAppointmentBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: ["string", "null"]

    AvailableSlotsResponse:
      type: object
      properties:
        date:
          type: string
        slots:
          type: array
          items:
            $ref: "#/components/schemas/TimeSlot"
      required:
        - date
        - slots

    TimeSlot:
      type: object
      properties:
        time:
          type: string
        available:
          type: boolean
      required:
        - time
        - available

    AppointmentStats:
      type: object
      properties:
        total:
          type: integer
        pending:
          type: integer
        confirmed:
          type: integer
        completed:
          type: integer
        cancelled:
          type: integer
        todayCount:
          type: integer
      required:
        - total
        - pending
        - confirmed
        - completed
        - cancelled
        - todayCount

    GalleryImage:
      type: object
      properties:
        id:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
      required:
        - id
        - imageUrl
        - caption
        - category
        - sortOrder
        - createdAt

    CreateGalleryImageBody:
      type: object
      properties:
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        sortOrder:
          type: integer
      required:
        - imageUrl

    Review:
      type: object
      properties:
        id:
          type: integer
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
        isApproved:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - customerName
        - rating
        - comment
        - isApproved
        - createdAt

    CreateReviewBody:
      type: object
      properties:
        customerName:
          type: string
        rating:
          type: integer
        comment:
          type: string
      required:
        - customerName
        - rating
        - comment

    ContactMessage:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - phone
        - message
        - isRead
        - createdAt

    CreateContactMessageBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        message:
          type: string
      required:
        - name
        - email
        - message

    AdminLoginBody:
      type: object
      properties:
        password:
          type: string
      required:
        - password

    AdminLoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    AdminMeResponse:
      type: object
      properties:
        authenticated:
          type: boolean
      required:
        - authenticated