# lib/api-zod/src/generated/types/appointmentStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Looks N Styles Beauty Salon API
 * OpenAPI spec version: 0.1.0
 */

export interface AppointmentStats {
  total: number;
  pending: number;
  confirmed: number;
  completed: number;
  cancelled: number;
  todayCount: number;
}