# lib/api-zod/src/generated/types/createAppointmentBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Looks N Styles Beauty Salon API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAppointmentBody {
  customerName: string;
  customerPhone: string;
  serviceId: number;
  appointmentDate: string;
  timeSlot: string;
  /** @nullable */
  notes?: string | null;
}