# lib/api-zod/src/generated/types/createContactMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Looks N Styles Beauty Salon API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateContactMessageBody {
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  message: string;
}